+++ server/game/session/ActionDispatcher.cpp
@@ -109,7 +109,7 @@ void ActionDispatcher::handleMove(uint32_t id, const Message &msg, GameWorld &wo
     if (!moving)
     {
         monitor.broadcast(std::make_shared<const EntityMoveMessage>(
-            static_cast<uint8_t>(id),
+            id,
             world.getPixelX(id),
             world.getPixelY(id),
             direction,
@@ -121,7 +121,7 @@ void ActionDispatcher::handleMove(uint32_t id, const Message &msg, GameWorld &wo
     if (world.movePlayer(id, direction))
     {
         monitor.broadcast(std::make_shared<const EntityMoveMessage>(
-            static_cast<uint8_t>(id),
+            id,
             world.getPixelX(id),
             world.getPixelY(id),
             direction,
